deleted old TransactionWindowController class
